requirements.txt
Django==3.0.7
gunicorn==19.9.0
Pillow==7.1.2
python-dotenv==0.13.0
django-model-utils==4.0.0
djangorestframework==3.11.0
celery==4.4.5
redis==3.5.3
django-celery-results==1.2.1
Markdown==3.2.2
numpy==1.18.5
sphinx==3.1.1
sphinx-rtd-theme==0.5.0
django-csp==3.6
psycopg2-binary==2.8.5
coverage==5.1
tblib==1.6.0
django-probes==1.5.0
oauthlib==3.1.0